+++ fiddles/vue/fiddle-0005-GoldenLayoutHwTs/src/views/DocsView.vue
@@ -1,0 +1,16 @@
+<script setup lang="ts">
+import TheWelcome from '../components/DocList.vue'
+</script>
+
+<template>
+  <main class="welcome-main">
+    <TheWelcome />
+  </main>
+</template>
+
+<style scoped>
+  .welcome-main {
+    padding: 100px;
+  }
+
+</style>