+++ fiddles/vue/fiddle-0006-GoldenLayoutHwTs/src/components/main.ts
@@ -1,0 +1,4 @@
+import GoldenLayout from "./GoldenLayout.vue";
+import TopNav from "./TopNav.vue";
+import BreadCrumbs from "./BreadCrumbs.vue";
+export { GoldenLayout, TopNav, BreadCrumbs };